pyproject.toml
[project]
name = "aiagentstudy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
